Add error metadata & tagging
This change adds support for `errors.New(...)` to receive metadata
and/or tags as arguments, and the means to introspect errors to get such
metadata and check for tag existence.

Tagging example:
    err := errors.New("oops: %s", errors.Tag("abort"), "arg")
    fmt.Println(err.Error()) // <- prints "oops: arg"
    fmt.Println(errors.HasTag("abort")) // <- prints "true"

Metadata example:
    err := errors.New("oops: %s", errors.Meta("user", "Joe"), "arg")
    fmt.Println(err.Error()) // <- prints "oops: arg"
    fmt.Println(errors.GetMeta(err, "user")) // <- prints "Joe"

Metadata and tag arguments are never incorporated into the error message
and are only available through the "Tags()" and "Meta()" functions.