Add function to print entire stack trace of error chain
Implemented a new function called PrintStackChain in the errors package.
This function prints the entire stack trace of an error chain to a given
writer. It's designed to provide more detailed error reporting than the
standard "%+v" format string in fmt.Printf, as it prints the entire
error chain even if there are wrapping errors in the chain that do not
normally print their wrapped errors. Tests have also been added to
ensure its correct functionality.