Fix chain stack-trace printing when it contains non-frames errors
This fixes an issue when, for an example chain of:

    FramesErr1 -> nonFramesErr -> FramesErr2

Where:
- "FramesErr1" is an error created by this library, and thus contains a
  stack-trace
- "nonFramesErr" is a simple error from an external library (or even the
  standard "errors" library)
- "FramesErr2", like the first, is also an error created by this library

In this example, printing the chain via "%+v" would print a partial
stack-trace, because the middle error, when formatted into printing,
will not continue printing the chain.

This commit fixes this issue by delegating to the "PrintStackChain"
function when an error is formatted by "%+v".